meta-adi-bsp: Add ADI bsp layer
This bsp layer is intended to be used with plain yocto and not with
petalinux as meta-adi-xilinx. As for now, there's only a recipe for
compiling ADI kernel and the available device tree's. ADI kernel is only
compiled for supported machines (as for now, zynq, zynqmp and microblaze)
which are defined via COMPATIBLE_MACHINE.This can be easily extended by
appending the recipe or directly changing it. Also note, that for building
for example for zc706-zynq7 machine, the meta-xilinx-bsp layer has to be added
since this layer is not redefining machines which are already defined in
another yocto layer's.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>