meta-adi-xilinx: dts: Do not overwrite system-user.dtsi
In the previous mechanism, the selected ADI devicetree was just being
copied and overwriting system-user.dtsi which is the file intended
to be used by the user when extending the core devicetree. So, with that
mechanism, it was impossible or very cumbersome to extend the ADI
devicetree being used. So that, now we just copy the selected dts to a
file named analog-devices.dtsi which is then included in the begining of
the system-user.dtsi file. This way, the user can freely customize the
system-user.dtsi file to his needs.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>