meta-adi-xilinx: linux: Refactor kernel recipe
Instead of just appending the xilinx kernel recipe which brings us some
issues as defining our ADI SRC_URI in an independent way, the adi kernel
recipe is added as new xilinx linux recipe version which is taken by default
by bitbake. On top of this, a linx-adi.inc file is introduced which
should hold all the common pieces for all the adi branches we want to
compile.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>