meta-adi-xilinx: devicetree: Enforce stdout-path node
This patch is adding another dtsi fragment for zynq platforms. In our
kernel devicetrees, for zynq, the stdout-path node is defined as
`stdout-path = "/amba@0/uart@E0001000"` which leads to

"
...
No serial driver found
resetting ...
...
"

when booting the device. For petalinux builds it seems we need to ensure
that `stdout-path = "serial0:115200n8"` so that uboot or fsbl can find a serial
device.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>