meta-adi-xilinx: Add hooks for FPGA Manager
This patch adds hooks to make it easier for someone to use the FPGA
Manager feature. In the end, is up to the user to provide a valid
devicetree overlay for the reference design being built. Furthermore,
with this patch the build won't fail if the FPGA Manager or the Devicetree
Overlays configs are enabled.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>