meta-adi-xilinx: device-tree: Refactor recipe
The motivation for this rework was to realize that the devicetrees
generated by petalinux based on our reference designs are not usable.
When using the petalinux devicetree, as is, and without any PL part (we
know already that we need to overwrite that part), we end up with a
system where, for example, no network devices are instantiated. So,
instead of just appending our devicetree to the petalinux one, resulting
in a lot of differences between the end devicetree and the ones we have
in our kernel tree, we completely overwrite it. However, we still
inlcude the petalinux pl.dtsi (so that user's can add IPs to our reference
designs and devicetree nodes are generated for those) and system-user.dtsi
(so that users can extend our devicetrees) in our top devicetree.
Like this, changes are kept to a minimal and sould only be introduced by
users and not by the build system.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>