device-tree. Support custom device-trees
Make it easier for users that want to use custom devicetrees. Until now,
users would either need to use or extend a supported devicetree. To support
a new one, the process was not that straight and some variables needed to
be changed.

This patch also removes the sanity check since it does not really make
sense to have it now. Moreover, we are already making sure that the
devicetree file exists and that should be enough...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>