device-tree: Force KERNEL_DTB to be defined out of the recipe
Defining configuration variables inside of the recipe is not really how it
should be done in yocto. With this patch, users are forced to configure the
devicetree to be used in their local.conf.

These could also be done using machine conf files for all the supported projects.
The reason why this was not done is that petalinux is generating the
machine configurations dynamically. While it seems the configurations are
pretty static (per platform family), we cannot really foresee all the corner
cases that can hit us badly...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>