meta-adi-xilinx: python: fix python for microblaze
Python is failing to build for microblaze with a similar error:

```
File "build/temp.linux2-microblazeel-2.7/libffi/fficonfig.py", line 33, in <module>
ffi_sources += ffi_platforms['MICROBLAZE']
KeyError: 'MICROBLAZE'
```

A patch [1] was once sent to OE to fix this issue to an older python
version. Naturally, the change introduced by the patch was already done
in the version we are trying to build. Hence, probably something changed
in the python build that is breaking microblaze. We may have to revisit
this, after getting some more info from the upstream community.

The patch being applied is fairly straight and it fixes the build. We
still need to check if enabling python bindings on libiio actually
works.

[1]: https://www.openembedded.org/pipermail/openembedded-core/2015-April/221941.html

Signed-off-by: Nuno Sa <nuno.sa@analog.com>