meta-adi-xilinx: u-boot: fix for microblaze
There are two problems with u-boot for microblaze:

1) As there are issues with microblaze and uartlite, the driver was not
   being selected in u-boot.

2) The auto generated platform header is likely not ok for ADI custom
   designs and we experience data bus exceptions on u-boot. It looks to be
   related with flash settings. So for now, as we don't really need it, we
   just disable mtd support on uboot and this fixes the abort.

This patch adds a configuration fragment to first force uartlite to be
compiled in and second to disable all mtd settings.

Also to mention that u-boot is needed since now initrd is the default
root on petalinux builds. Thus, we use u-boot to load the image.ub into
memory and then boot from it...

Signed-off-by: Nuno Sa <nuno.sa@analog.com>