meta-adi-xilinx: fsboot: fix uartlite selection
There's an issue in petalinux for microblaze designs using the uartlite
core. This change forces the fsboot stdin/stdout to be axi_uart. If
someone is using a custom design with a different core with a different
name, will need to change `FSBOOT_CONSOLE_STDIN` and
`FSBOOT_CONSOLE_STDOUT` accordingly in a conf file.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>