meta-adi-xilinx: devicetree: tweak system-conf.dtsi
The only node we care about in system-conf is the chosen node. The other
ones are likely to "collide" with nodes already defined in our own
devicetrees. For that reason, we were just deleting these nodes.
However, the list of nodes to delete in the sed command was starting to
grow, so that, with this patch we change the approach and we tweak the
file to delete everything but the chosen node.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>