meta-adi-xilinx: devicetree: fix uartlite selection
There's an issue in petalinux for microblaze designs using the uartlite
core. That means, that we need to force it so that xilinx tools do not
complain that there's no serial selected. The solution (and for more
info) was taken from here [1].

If someone is using a custom design with a different core with a different
name, will need to change `DTS_CONSOLE_DEVICE_CONFIG` accordingly in a conf
file.

[1]: https://forums.xilinx.com/t5/Embedded-Linux/AXI-Uartlite-issues-with-Petalinux-2020-1-and-Microblaze-systems/td-p/1126614
Signed-off-by: Nuno Sá <nuno.sa@analog.com>