meta-adi-xilinx: devicetree: copy all devicetrees to workdir
This change will copy all the devicetrees (being in kernel or not) to
the WORKDIR directory. The goal is to apply our pl + pl-delete-nodes
logic on the $dtb_tag_file without directly changing the original file
In case it's an in kernel devicetree, changing it would actually
break the kernel compilation

With petalinux 2020.1, in case it's an in kernel devicetree, changing it
would actually break the kernel compilation since compiling the kernel
will also compile all the devicetrees by default.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>