meta-adi-xilinx: libiio: fix branch selection
Add the sysvinit patch in the same SRC_URI append line. Otherwise, the
'+=' operator has precedence over '_append' which means that the ';branch='
key will not be applied to the 'git://' URI and hence, we cannot specify
the branch to checkout. It will always use master no matter what is
given in BRANCH'.

Fixes: cd042513d600 ("meta-adi-xilinx: libiio: remove lsb runtime dependency")
Signed-off-by: Nuno Sá <nuno.sa@analog.com>