meta-adi-xilinx: conf: increase priority to 8
This is needed because the default priority of meta-user in petalinux
was increased to 7. Hence, the 'do_configure_append' task present in the
devicetree.bbappend of the meta-user layer would only run after the
append task present in meta-adi-xilinx. This would lead to a build
failure because system-conf.dtsi is generated by the meta-user layer.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>