device-tree: Fix issue for custom DTS
If someone uses the recommendations from the README.md file
to create a custom device tree, the build runs into an error,
because 'cp' uses the same values for source and destination
if "KERNEL_DTB_PATH" is set to "workdir". This results in a return
value unequal to zero and stops executing the build process.

Moreover, the string comparison from USE_KERNEL_SOURCES did not
contain double quotes, thus it was not interpreted as a string.

Fixes: a0267decd ("meta-adi-xilinx: devicetree: copy all devicetrees to workdir")

Signed-off-by: Andre Werner <andre.werner@systec-electronic.com>