meta-adi-xilinx: petalinux-image-minimal: add analog user
This change pretty much copies what default meta-petalinux is doing:
 * Add a a new user;
 * Add that user to the audio and video group;
 * Add the aie group and the user to it;
 * Add the user to sudoers.

The differences are that the username is analog instead of petalinux and
we hardcode the root password to 'analog' so that we have a closer
experience with what we have in kuiper.

As before, changing the password with 'petalinux-config -c rootfs' won't
work. We might re-think and improve it in the future.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>