meta-adi-xilinx: device-tree: make sure DTS_INCLUDE_PATH is defined
If compiling for a MACHINE other than zynq, zynqMP or microblaze based,
DTS_INCLUDE_PATH would evaluate to nothing since there was not valid
OVERRIDE for it. Hence, this would result in copying the complete root
(/) directory into the build system. This was happening when trying to
build for versal platforms.

Instead of just adding another override for versal, redefine how things are
done by having a default value pointing to the "base" devicetree path.
Then, fine tune with OVERRIDES when needed. For example, for zynqMP we
only care about the xilinx devicetree directory. With this, worst case
scenario, we just end up copying some more things than needed (complete
dts folder) if the MACHINE is not supported by meta-adi.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>