meta-adi-xilinx: recipes-support: libiio: move to libbio-v0 branch
Libiio renamed the master branch to main. More importantly, it moved to
libiio v1.0 which is not backward compatible (at least for now) with
libiio v0. Hence, all the projects that 'DEPEND' on libiio will fail to
compile in meta-adi. Since it might take some time for those projects to
be converted, let's move to the v0 branch and update the package 'PV'
accordingly.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>