meta-adi-xilinx: petalinux-image-minimal: remove fsboot
Fsboot fails to build in vc707 and kc705 based projects. This is a
petalinux bug or something missing in our hdl projects that petalinux is
expecting to exist. Anyways,  we never really used fsboot as part of our
workflow and, AFAIK, no one is really using it. Hence just don't build it.
If someone raises his hand, we can then dig further.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>