meta-adi-core: remove layer
All the recipes built into meta-adi-core were moved directly into
meta-adi-xilinx. The idea of the core layer was that it could be used
by different ADI layers. Well, in all meta-adi existence, that was
never needed/happened. Moreover it actually brings some strange issues
with petalinux (more visible in 2023.1). The problem is that there's no
guaranteed order in which USER layers are added which means that if
there are dependencies between the layers (as it was the case with
meta-adi-xilinx and meta-adi-core), `petalinux-conf` will fail.

While at it, yocto is now issuing warnings if the licenses don't use
SPDX form. Hence, let's update the licenses string while moving the
files.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>