meta-adi-xilinx: petalinux-image-minimal: really remove avahi-daemon
On Microblaze, zeroconf is not enabled so that we can remove
avahi-daemon from the image. However, we were trying to do the remove
action after the microblaze override and so, the package was still
included in the image. Fix it by placing the override in it's place
after the ':remove' action.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>