meta-adi-xilinx: devicetree: copy dtc/include-prefixes
Also copy dtc/include-prefixes into the WORKDIR. In 2023_R2, xilinx is
clearing (using the ?= operator) KERNEL_INCLUDE which means any
devicetree including an header in dt-bindings, for example, would fail
to build. We could just change our own assignment from '?=' to '=' so we
would force our value but instead of messing with a variable used by the
devicetree class and petalinux, let's just copy the complete directory
into WORKDIR. As this is also something that lives in the kernel, we
also do that copy if 'USE_KERNEL_SOURCES' is set.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>