ci: petalinux-build: update build script
The script now will be maintained per branch. That means that there's no
need to make sure it works for more than one petalinux version as now it
"knows" what version it must run on.

On top of this, we now use DL_DIR to speed up a bit the build.

Also note that the script on the server was also changed. Now, it's a
simple wrapper that only checks which meta-adi branch is needed and then
run this script.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>