meta-adi-xilinx: README.md: Update for 2023_R2
Update all links to point to 2023_R2 branches. Also marked some projects
as obsolete. These are kept until they exist in the hdl tree so that
users might still use them (though it is not guaranteed they build).

Note that some links that refer to meta-adi 2023_R2 based files will be
temporarily broken as we'll only branch out once the 2023 update is done
on main.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>