meta-adi-xilinx: linux: fix recipe selection
The recipe being used to build the kernel was the 2023.1 recipe. Since
we append the xilinx recipe and "force brute" ADI kernel and version, any
recipe would match PREFERRED_VERSION. Hence, bitbake selects the first
available recipe. To fix this take the version extension from the recipe
filename version and match that in PREFERRED_VERSION. This is taken from
Xilinx base recipe.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>