meta-adi-xilinx: README: explanations on disabling analog user
The section about the default Kuiper Linux behaviour used to propose
commenting out lines in petalinux-image-minimal.bbappend of this
project in order to disable the default behaviour. This is a bad
practice since it breaks layered design of yocto project. The new way
of doing it is to add the 2 following lines to petalinuxbsp.conf
of user layer, meta-user, instead:
KUIPER_COMPAT_SUDOERS = ""
KUIPER_COMPAT_USERADD = ""

Signed-off-by: sumguytho <sumguytho@gmail.com>